+++ tests/Behat/Bootstrap/UserContextTrait.php
@@ -10,6 +10,26 @@ use Redmine\Api\User;
 trait UserContextTrait
 {
     /**
+     * @Given I create :count users
+     */
+    public function iCreateUsers(int $count)
+    {
+        while ($count > 0) {
+            $table = new TableNode([
+                ['property', 'value'],
+                ['login', 'testuser_' . $count],
+                ['firstname', 'first'],
+                ['lastname', 'last'],
+                ['mail', 'mail.' . $count . '@example.net'],
+            ]);
+
+            $this->iCreateAUserWithTheFollowingData($table);
+
+            $count--;
+        }
+    }
+
+    /**
      * @When I create a user with the following data
      */
     public function iCreateAUserWithTheFollowingData(TableNode $table)
@@ -53,6 +73,20 @@ trait UserContextTrait
 
         $this->registerClientResponse(
             $api->list(),
+            $api->getLastResponse(),
+        );
+    }
+
+    /**
+     * @When I list all user logins
+     */
+    public function iListAllUserLogins()
+    {
+        /** @var User */
+        $api = $this->getNativeCurlClient()->getApi('user');
+
+        $this->registerClientResponse(
+            $api->listLogins(),
             $api->getLastResponse(),
         );
     }